fix WaitAll Bug. (but still not stable)
